Change @isTestGroup to @isTest on testGoldens
This method represents a test, not a group of tests. This annotation is used by editors like VS Code to locate tests/groups and merges them with the test results in the tree. Using `isTestGroup` would cause the analysis-based test discovery to add a group with this name, whereas the test results would add a test with this name. Both should be the same.