The "physical size" should be "logical size" * "pixel ratio"
In order to actually have the golden image resolution match the expected physical size, the physical size needs to be set to the logical size * pixel ratio. Before this, an image with size: Size(100.0, 100.0) and devicePixelRatio: 2.0 would result in a .png of size 100 x 100, while we'd expect to have an image of 200 x 200 pixels.